+++ src/bindings/GithubSlugger.js
