+++ src/components/SyntaxLookupWidget.js
@@ -1,15 +1,15 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Icon from "./Icon.js";
+import * as Next from "../bindings/Next.js";
 import * as Curry from "bs-platform/lib/es6/curry.js";
 import * as React from "react";
 import * as Js_dict from "bs-platform/lib/es6/js_dict.js";
 import FuseJs from "fuse.js";
 import * as Markdown from "./Markdown.js";
+import * as SearchBox from "./SearchBox.js";
 import * as Belt_Array from "bs-platform/lib/es6/belt_Array.js";
 import * as Belt_Option from "bs-platform/lib/es6/belt_Option.js";
 import * as Caml_option from "bs-platform/lib/es6/caml_option.js";
-import * as Router from "next/router";
 import * as GithubSlugger from "github-slugger";
 import Decorator_asMdx from "misc_docs/syntax/decorator_as.mdx";
 import Decorator_moduleMdx from "misc_docs/syntax/decorator_module.mdx";
@@ -100,100 +100,6 @@ function getAnchor(path) {
   }
 }
 
-function SyntaxLookupWidget$SearchBox(Props) {
-  var completionValuesOpt = Props.completionValues;
-  var value = Props.value;
-  var onClear = Props.onClear;
-  var onValueChange = Props.onValueChange;
-  var completionValues = completionValuesOpt !== undefined ? completionValuesOpt : [];
-  var match = React.useState(function () {
-        return /* Inactive */1;
-      });
-  var setState = match[1];
-  var state = match[0];
-  var textInput = React.useRef(null);
-  var onMouseDownClear = function (evt) {
-    evt.preventDefault();
-    return Curry._1(onClear, undefined);
-  };
-  var onAreaFocus = function (evt) {
-    var el = evt.target;
-    var isDiv = (el.type == null);
-    if (isDiv && state === /* Inactive */1) {
-      return Belt_Option.forEach(Caml_option.nullable_to_opt(textInput.current), (function (el) {
-                    el.focus();
-                    
-                  }));
-    }
-    
-  };
-  var onFocus = function (param) {
-    return Curry._1(setState, (function (param) {
-                  return /* Active */0;
-                }));
-  };
-  var onBlur = function (param) {
-    return Curry._1(setState, (function (param) {
-                  return /* Inactive */1;
-                }));
-  };
-  var onKeyDown = function (evt) {
-    var key = evt.key;
-    var ctrlKey = evt.ctrlKey;
-    var full = (
-      ctrlKey ? "CTRL+" : ""
-    ) + key;
-    switch (full) {
-      case "Escape" :
-          return Curry._1(onClear, undefined);
-      case "Tab" :
-          if (completionValues.length !== 1) {
-            return ;
-          }
-          var targetValue = Belt_Array.getExn(completionValues, 0);
-          if (targetValue !== value) {
-            evt.preventDefault();
-            return Curry._1(onValueChange, targetValue);
-          } else {
-            return ;
-          }
-      default:
-        return ;
-    }
-  };
-  var onChange = function (evt) {
-    evt.preventDefault();
-    return Curry._1(onValueChange, evt.target.value);
-  };
-  return React.createElement("div", {
-              className: (
-                state === /* Active */0 ? "border-fire" : "border-fire-40"
-              ) + " flex items-center border rounded-lg py-4 px-5",
-              tabIndex: -1,
-              onFocus: onAreaFocus,
-              onBlur: onBlur
-            }, React.createElement(Icon.MagnifierGlass.make, {
-                  className: (
-                    state === /* Active */0 ? "text-fire" : "text-fire-80"
-                  ) + " w-4 h-4"
-                }), React.createElement("input", {
-                  ref: textInput,
-                  className: "text-16 outline-none ml-4 w-full",
-                  placeholder: "Enter keywords or syntax...",
-                  type: "text",
-                  value: value,
-                  onKeyDown: onKeyDown,
-                  onFocus: onFocus,
-                  onChange: onChange
-                }), React.createElement("button", {
-                  className: value === "" ? "hidden" : "block",
-                  onFocus: onFocus,
-                  onMouseDown: onMouseDownClear
-                }, React.createElement(Icon.Close.make, {
-                      className: "w-4 h-4 text-fire"
-                    })));
-}
-
 function SyntaxLookupWidget$Tag(Props) {
   var text = Props.text;
   return React.createElement("span", {
@@ -235,7 +141,7 @@ function SyntaxLookupWidget$DetailBox(Props) {
 }
 
 function SyntaxLookupWidget(Props) {
-  var router = Router.useRouter();
+  var router = Next.Router.useRouter(undefined);
   var match = React.useState(function () {
         return /* ShowAll */0;
       });
@@ -261,23 +167,23 @@ function SyntaxLookupWidget(Props) {
   React.useEffect((function () {
           var match = getAnchor(router.asPath);
           var exit = 0;
-          if (typeof state === "number" || !state.TAG) {
+          if (typeof state === "number" || state.TAG === /* ShowFiltered */0) {
             exit = 1;
           } else {
             var item = state._0;
             if (match !== undefined) {
               var slug = GithubSlugger.slug(item.name);
               if (slug !== match) {
-                router.replace("syntax-lookup#" + match);
+                Next.Router.replace(router, "syntax-lookup#" + match);
               }
               
             } else {
-              router.replace("syntax-lookup#" + GithubSlugger.slug(item.name));
+              Next.Router.replace(router, "syntax-lookup#" + GithubSlugger.slug(item.name));
             }
           }
           if (exit === 1) {
             if (match !== undefined) {
-              router.replace("syntax-lookup");
+              Next.Router.replace(router, "syntax-lookup");
             }
             
           }
@@ -317,7 +223,7 @@ function SyntaxLookupWidget(Props) {
                 }));
   };
   var details;
-  if (typeof state === "number" || !state.TAG) {
+  if (typeof state === "number" || state.TAG === /* ShowFiltered */0) {
     details = null;
   } else {
     var item = state._0;
@@ -341,7 +247,7 @@ function SyntaxLookupWidget(Props) {
         }));
   var items;
   items = typeof state === "number" ? allItems : (
-      state.TAG ? [] : state._1
+      state.TAG === /* ShowFiltered */0 ? state._1 : []
     );
   var categories = Belt_Array.reduce(Js_dict.entries(Belt_Array.reduce(items, Js_dict.fromArray(initial), (function (acc, item) {
                   var key = toString(item.category);
@@ -391,16 +297,16 @@ function SyntaxLookupWidget(Props) {
       "",
       allItems
     ];
-  } else if (state.TAG) {
+  } else if (state.TAG === /* ShowFiltered */0) {
+    match$1 = [
+      state._0,
+      state._1
+    ];
+  } else {
     var item$1 = state._0;
     match$1 = [
       item$1.name,
       [item$1]
-    ];
-  } else {
-    match$1 = [
-      state._0,
-      state._1
     ];
   }
   var onSearchClear = function (param) {
@@ -424,12 +330,13 @@ function SyntaxLookupWidget(Props) {
                       style: {
                         maxWidth: "34rem"
                       }
-                    }, React.createElement(SyntaxLookupWidget$SearchBox, {
+                    }, React.createElement(SearchBox.make, {
                           completionValues: Belt_Array.map(match$1[1], (function (item) {
                                   return item.name;
                                 })),
                           value: match$1[0],
                           onClear: onSearchClear,
+                          placeholder: "Enter keywords or syntax...",
                           onValueChange: onSearchValueChange
                         }))), React.createElement("div", {
                   className: "mt-10"
