+++ src/components/SyntaxLookupWidget.resi
